+++ Tests/Helpers/Io_Xpring_SubmitSignedTransactionResponse+Test.swift
@@ -1,7 +1,0 @@
-import XpringKit
-
-extension Io_Xpring_SubmitSignedTransactionResponse {
-  static let testSubmitTransactionResponse = Io_Xpring_SubmitSignedTransactionResponse.with {
-    $0.transactionBlob = .testTransactionBlobHex
-  }
-}